api/requirements.txt
python-dotenv
fastapi
httpx
uvicorn
spotipy